+++ src/java/agendaalineweb/controllers/ExcluirCliente.java
@@ -5,6 +5,7 @@
 package agendaalineweb.controllers;
 
 import agendaalineweb.entities.Cliente;
+import agendaalineweb.entities.Usuario;
 import agendaalineweb.models.AgendamentoModel;
 import agendaalineweb.models.ClienteModel;
 import java.io.IOException;
@@ -22,33 +23,43 @@ import java.util.ArrayList;
  */
 @WebServlet(name = "ExcluirCliente", urlPatterns = {"/excluir-cliente"})
 public class ExcluirCliente extends HttpServlet {
-    
-    
+
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        String id = request.getParameter("id");
-        int idConvertido = Integer.parseInt(id);
-        ClienteModel modelCliente = new ClienteModel();
-         String mensagem = null;
-         AgendamentoModel agendamentoModel = new AgendamentoModel();
-        boolean clienteTemAgendamento = agendamentoModel.clienteTemAgendamento(idConvertido);
-        if(clienteTemAgendamento){
-            mensagem = "O cliente selecionado não pode ser excluido. Há agendamentos associados a este cliente";
-        }else{
-            modelCliente.deleteById(idConvertido);
+        request.setCharacterEncoding("UTF-8");
+        response.setContentType("text/html;charset=UTF-8");
+        String mensagem = null;
+         ClienteModel modelCliente = new ClienteModel();
+        Usuario usuario = (Usuario) request.getSession().getAttribute("usuarioLogado");
+        if (usuario != null) {
+
+            String id = request.getParameter("id");
+            int idConvertido = Integer.parseInt(id);
+           
+
+            AgendamentoModel agendamentoModel = new AgendamentoModel();
+            boolean clienteTemAgendamento = agendamentoModel.clienteTemAgendamento(idConvertido);
+            if (clienteTemAgendamento) {
+                mensagem = "O cliente selecionado não pode ser excluido. Há agendamentos associados a este cliente.";
+            } else {
+                modelCliente.deleteById(idConvertido);
+            }
+            
+        } else {
+            mensagem = "Você deve estar logado para executar essa operação.";
         }
-        ArrayList<Cliente> clientes = modelCliente.selectAll();
-        request.setAttribute("clientes", clientes);
-        request.setAttribute("mensagemErro", mensagem);
-        String caminhoContexto = request.getContextPath();
-        request.setAttribute("caminhoContexto", caminhoContexto);
-        request.getRequestDispatcher("WEB-INF/pageCliente.jsp").forward(request, response);
+        ArrayList<Cliente> clientes = modelCliente.selectAll(usuario.getIdNegocio());
+            request.setAttribute("clientes", clientes);
+            request.setAttribute("mensagemErro", mensagem);
+            String caminhoContexto = request.getContextPath();
+            request.setAttribute("caminhoContexto", caminhoContexto);
+            request.getRequestDispatcher("WEB-INF/pageCliente.jsp").forward(request, response);
     }
-    
+
     @Override
     public String getServletInfo() {
         return "Short description";
     }
-    
+
 }
